+++ modules/init.js
@@ -1,5 +1,3 @@
-import VersionCheck from "./versioning/version-check.mjs";
-import renderWelcomeScreen from "./versioning/welcome-screen.mjs";
 import constants from "./constants.mjs";
 import registerDerivedItemSheetClass from "./ItemSheet.js";
 import registerContextMenuHook from "./ContextMenu.mjs";
@@ -22,12 +20,6 @@ Hooks.once('setup', () => {
 });
 
 Hooks.once("ready", () => {
-  if (VersionCheck.check(constants.moduleName)) {
-    if (game.user.isGM || game.settings.get(constants.moduleName, 'playersWelcomeScreen')) {
-      renderWelcomeScreen();
-    }
-  }
-
   checkSettingsInitialized();
   registerDerivedItemSheetClass();
   registerItemClassMethod();